made geoips test workflow accept unregistered workflows
