+++ geoips/commandline/geoips_command.py
@@ -8,22 +8,28 @@ Will implement a plethora of commands, but for the meantime, we'll work on
 'geoips validate'.
 """
 
-import os
 import abc
 import argparse
-import warnings
+from ast import literal_eval
+from collections.abc import Mapping
 from importlib import metadata, resources
 import json
+import os
+from pathlib import Path
 from shutil import get_terminal_size
+from typing import List, Union
+import warnings
 
 from colorama import Fore, Style
 from pluginify.config import REGISTRY_DIRECTORY
 from tabulate import tabulate
 
-import geoips.utils.yaml_utils as yaml
 from geoips.commandline.ancillary_info import cmd_instructions, alias_mapping
 from geoips.commandline.log_setup import setup_logging
 from geoips.filenames.base_paths import PATHS
+from geoips.interfaces import workflows
+from geoips.pydantic_models.v1.workflows import WorkflowPluginModel
+import geoips.utils.yaml_utils as yaml
 
 
 class PluginPackages:
@@ -435,7 +441,7 @@ class GeoipsExecutableCommand(GeoipsCommand):
     def __init__(self, LOG, parent=None, legacy=False):
         """Initialize GeoipsExecutableCommand.
 
-        This is a child of GeoipsCommand and will invoke the functionaly of
+        This is a child of GeoipsCommand and will invoke the functionality of
         GeoipsCommand __init__ func alongside additional logic needed to set up
         executable-based commands. This will instantiate each subcommand class with a
         parser and point towards the correct default function to call if that subcommand
@@ -460,7 +466,7 @@ class GeoipsExecutableCommand(GeoipsCommand):
               suppressing or displaying help information for '--procflow'.
         """
         super().__init__(LOG=LOG, parent=parent, legacy=legacy)
-        # Since this class is exectuable (ie. not the cli, top-level list...),
+        # Since this class is executable (ie. not the cli, top-level list...),
         # add available arguments for that command and set that function to
         # the command's executable function (__call__) if that command is called.
         self.add_arguments()
@@ -746,6 +752,119 @@ class GeoipsExecutableCommand(GeoipsCommand):
                     plugin_entry.append(plugin_dict[plugin_key][header])
             table_data.append(plugin_entry)
         return table_data
+
+
+class GeoipsWorkflowCommand(GeoipsExecutableCommand):
+    """Executable command class used for two workflow-specific commands.
+
+    Those commands include ``geoips test workflow`` and ``geoips run order_based``.
+    This class has been created to share functionality for detecting input workflows
+    and casting them to the validating that their input matches one of three types.
+
+    Accepted inputs include the name of a registered workflow plugin, a .json or .yaml
+    filepath to an unregistered workflow plugin, or a dictionary representing a
+    workflow that will be literally evaluated as such.
+    """
+
+    def ensure_valid_json_or_yaml_path(self, value: str) -> Path:
+        """Ensure 'value' is a valid path to a json/yaml file and convert to a Path object.  # NOQA
+
+        Parameters
+        ----------
+        value: str
+            - The input value for the filepath to typecheck against.
+
+        Returns
+        -------
+        path: Path
+            - A json or yaml pathlib.Path object.
+        """
+        path = Path(value)
+
+        if path.suffix.lower() not in {".json", ".yaml", ".yml"}:
+            raise self.parser.error(
+                f"File must have extension .json, .yaml, or .yml, not {path.suffix}"
+            )
+
+        if not path.exists():
+            raise self.parser.error(f"Input filepath not found: {value}")
+
+        return path
+
+    def workflow_type(self, value: Union[str, List[Path], dict]):
+        """Cast input value to a workflow type.
+
+        If value cannot be cast to an accepted workflow type, argparse will raise an
+        error denoting that your argument value could not be associated with an
+        accepted type.
+
+        Parameters
+        ----------
+        value: Union[str, List[Path], dict]
+            - The input value of a potential workflow.
+
+        Returns
+        -------
+        workflow: WorkflowPlugin-like
+            - The same input value as long as it could be cast to any of the accepted
+              workflow types. Workflow has been automatically expanded in the case that
+              we need to apply overrides to it. Doesn't change the functionality of the
+              OBP if overrides don't occur.
+        """
+        # unregistered generated workflow
+        try:
+            workflow = literal_eval(value)
+        except (ValueError, SyntaxError):
+            # Ignore these errors, could still be valid input
+            workflow = None
+
+        if isinstance(workflow, Mapping):
+            # Validate the generated workflow with is_registered set to false as this
+            # plugin has been dynamically generated
+            workflow = WorkflowPluginModel(
+                **workflow,
+                is_registered=False,
+                # Adding context in pydantic is akin to passing in values that are
+                # usually None to an Objects __init__ function. It will construct
+                # differently if those parameters are provided. In this case, we are
+                # telling pydantic to expand the workflow, rather than validate just
+                # what's in the data provided
+                context={"expand": True},
+            ).model_dump()
+        # registered named workflow
+        elif isinstance(value, str):
+            workflow = workflows.get_plugin(value, _expand=True)
+        # unregistered workflow @ filepath
+        elif self.ensure_valid_json_or_yaml_path(value):
+            # since the filepath was valid and exists, load the data and validate it
+            filepath = self.ensure_valid_json_or_yaml_path(value)
+            if filepath.suffix.lower() == ".json":
+                loader = json.load
+            else:
+                loader = yaml.safe_load
+
+            with open(filepath, "r") as f:
+                workflow = loader(f)
+            # This assumes if you pass the filepath option that the plugin itself is not
+            # registered. Validate that it's formatted correctly.
+            workflow = WorkflowPluginModel(
+                **workflow,
+                is_registered=False,
+                # Adding context in pydantic is akin to passing in values that are
+                # usually None to an Objects __init__ function. It will construct
+                # differently if those parameters are provided. In this case, we are
+                # telling pydantic to expand the workflow, rather than validate just
+                # what's in the data provided
+                context={"expand": True},
+            ).model_dump()
+        else:
+            self.parser.error(
+                "Error: positional argument 'workflow' could not be associated with an"
+                f" accepted type. Input = {value} ; accepted types = "
+                "[str, Path, dict]"
+            )
+
+        return workflow
 
 
 class CommandClassFactory:
